backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
pyjwt[crypto]
passlib[bcrypt]
python-dotenv
